-stars in the promenade done =>optimization with star trail =>script 2 (OBSERVE) can now focus on Element, and just Perso =>new trigger : USE

+++ zildo/src/zildo/monde/sprites/persos/ia/Mover.java
@@ -78,7 +78,7 @@ public abstract class Mover {
 				entity.y += delta.y;
 			}
 		}
-		if (mobile.x == target.x && mobile.y == target.y) {
+		if ((int) mobile.x == target.x && (int) mobile.y == target.y) {
 			// Mover has accomplished his duty
 			active = false;
 		}
